:lipstick: style: 주문/포장 선택 화면 수정  (#63)
