+++ src/components/layout/order/footer/styles/Footer.module.css
@@ -1,3 +1,5 @@
 .container {
+    position: fixed;
     height: var(--kiosk-footer-height);
+    bottom: 0;
 }
