+++ src/components/order/order-type/styles/ButtonContainer.module.css
@@ -1,13 +1,5 @@
 .container {
     display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-}
-
-.container > div {
-    width: 100%;
-    display: flex;
-    justify-content: space-around;
+    align-items: flex-start;
+    gap: 20px;
 }
