+++ src/components/order/order-type/styles/OrderTypeButton.module.css
@@ -1,7 +1,35 @@
-.button {
-    border-width: 1px;
-    border-radius: .75rem;
-    width: 40%;
-    height: 10rem;
-    cursor: pointer;
+.wrapper {
+    display: flex;
+    width: 12rem  /* 120/10 */;
+    height: 15rem  /* 150/10 */;
+    padding: 2rem  /* 20/10 */ 1rem  /* 10/10 */;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 3rem  /* 30/10 */;
+
+    border-radius: 1rem  /* 10/10 */;
+    background: #FFF;
+    box-shadow: 0 0 1rem 0 rgba(0, 0, 0, 0.10);
+}
+
+.wrapper img {
+    display: flex;
+    width: 6rem  /* 60/10 */;
+    height: 6rem  /* 60/10 */;
+    padding: 1rem  /* 10/10 */;
+    justify-content: center;
+    align-items: center;
+    gap: 1rem  /* 10/10 */;
+    flex-shrink: 0;
+    border-radius: 5rem  /* 50/10 */;
+}
+
+.wrapper span {
+    color: #000;
+    font-family: Pretendard, sans-serif;
+    font-size: 12px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 12px; /* 100% */
 }
