+++ src/components/order/order-type/styles/OrderTypeContainer.module.css
@@ -4,7 +4,42 @@
     height: 100vh;
 }
 
+.choose-wrapper {
+    width: 100%;
+    display: inline-flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 5rem  /* 50/10 */;
+    margin: 50% 6rem /* 60/10 */;
+}
+
 .container h1 {
-    position: absolute;
-    top: 25%;
+    color: #000;
+    font-family: Pretendard, sans-serif;
+    font-size: 2rem  /* 20/10 */;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 2rem  /* 20/10 */; /* 100% */
+}
+
+.language-wrapper {
+    position: fixed;
+    bottom: 20px;
+    left: 20px;
+    display: inline-flex;
+    align-items: center;
+    gap: 10px;
+}
+.language-wrapper img {
+    width: 23px;
+    height: 23px;
+}
+
+.language-wrapper span {
+    color: var(--main-color);
+    font-family: Pretendard, sans-serif;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 14px; /* 100% */
 }
